+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Added `/filesystem/{system_name}/transfer/compress` and `/filesystem/{system_name}/transfer/extract`
+  - `compress` operations (on `transfer` and `ops` endpoints) accept `pattern` parameter to compress files using `regex` syntax.
 
 ### Changed